requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.1
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.7
Faker==22.2.0
flake8==7.0.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.4
mccabe==0.7.0
openapi-codec==1.3.2
packaging==23.2
pipreqs==0.4.13
pluggy==1.3.0
pycodestyle==2.11.1
pycountry==23.12.11
pyflakes==3.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.0.3
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
yarg==0.1.9
